requirements.txt
streamlit==1.22.0
opencv-python-headless==4.7.0.72
torch==2.1.0
ultralytics==8.0.100
Pillow==10.0.0
numpy==1.25.2
omegaconf==2.3.0

